+++ src/main/java/de/pixelmindmc/pixelchat/constants/LangConstants.java
@@ -34,6 +34,7 @@ public class LangConstants {
     public static final String PIXELCHAT_REPORT_BUGS = "pixelchat.report-bugs";
     public static final String PIXELCHAT_RELOAD = "pixelchat.reload";
     public static final String PIXELCHAT_REMOVED_PLAYER_STRIKES = "pixelchat.removed-player-strikes";
+    public static final String PIXELCHAT_STRUCK_PLAYER = "pixelchat.struck-player";
     // Pixelchat Guardian
     public static final String CLEARED_STRIKES_ON_SERVER_RESTART = "cleared-strikes-on-server-restart";
     public static final String PLAYER_MESSAGE_BLOCKED = "player.message-blocked";