+++ src/main/java/de/pixelmindmc/pixelchat/utils/TabCompleter.java
@@ -41,8 +41,12 @@ public class TabCompleter implements org.bukkit.command.TabCompleter {
                 if (sender.hasPermission(PermissionConstants.PIXELCHAT_RELOAD)) results.add("reload");
                 if (sender.hasPermission(PermissionConstants.PIXELCHAT_REMOVE_PLAYER_STRIKES))
                     results.add("remove-strikes");
-            } else if (args.length == 2 && args[0].equals("remove-strikes") || args[0].equals("removestrikes") || args[0].equals("rmstrikes"))
-                addOnlinePlayerCompletions();
+                if (sender.hasPermission(PermissionConstants.PIXELCHAT_STRIKE_PLAYER)) results.add("strike");
+            } else if (args.length == 2) {
+                if (args[0].equals("remove-strikes") || args[0].equals("removestrikes") || args[0].equals("rmstrikes"))
+                    addOnlinePlayerCompletions();
+                if (args[0].equals("strike")) addOnlinePlayerCompletions();
+            }
         }
 
         return results;