+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [1.1.0] - TBD
 
+### Added
+
 - Add support for the `CarbonChat` plugin
 - Add custom ChatGuard prefix
 - Add color & format code system similar to emoji system
+- Add strike command
+
+### Enhanced
+
+- Strike system and saving
 
 ## [1.0.1] - Latest Stable Release (11.10.2024)